refactor: extract FormatEntity format helpers and fix container/codec labeling
Move codecLabel(), containerLabel(), formattedBitrate(), formattedSampleRate(),
and formattedFileSize() into FormatEntity.kt as public extensions,
replacing ~70 lines of duplicated inline parsing scattered across 6 queue composables
and the private codecLabel() in PlayerComponents.kt.

The inline parsing in QueueComponents named its variable 'codec' while actually
extracting the container from mimeType (e.g. 'WEBM' instead of 'opus').
Variable is now honestly named 'container'. User-facing display is unchanged.

Bonus: bitrate formatting now returns 'Unknown' instead of '0 kbps' for local
files where bitrate is not available.
