+++ app/src/main/kotlin/moe/rukamori/archivetune/db/entities/FormatEntity.kt
@@ -9,6 +9,7 @@ package moe.rukamori.archivetune.db.entities
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import kotlin.math.roundToInt
 
 @Entity(tableName = "format")
 data class FormatEntity(
@@ -23,3 +24,34 @@ data class FormatEntity(
     val perceptualLoudnessDb: Double? = null,
     val playbackUrl: String?
 )
+
+fun FormatEntity.containerLabel(): String =
+    mimeType.substringAfter("/").substringBefore(";").uppercase()
+
+fun FormatEntity.codecLabel(): String {
+    val rawCodec = codecs.ifBlank { mimeType.substringAfter("/") }.uppercase()
+    val rawMime = mimeType.substringAfter("/").substringBefore(";").uppercase()
+
+    return when {
+        rawCodec.contains("FLAC") || rawCodec.contains("ALAC") -> "Lossless"
+        rawCodec.contains("OPUS") -> "OPUS"
+        rawCodec.contains("AAC") || rawCodec.contains("MP4A") -> "AAC"
+        rawCodec.contains("VORBIS") -> "VORBIS"
+        rawMime.contains("OPUS") -> "OPUS"
+        rawMime.contains("AAC") || rawMime.contains("MP4A") -> "AAC"
+        rawMime.contains("VORBIS") -> "VORBIS"
+        rawMime.isNotBlank() -> rawMime
+        else -> rawCodec
+    }
+}
+
+fun FormatEntity.formattedBitrate(): String =
+    if (bitrate > 0) "${bitrate / 1000} kbps" else "Unknown"
+
+fun FormatEntity.formattedSampleRate(): String? =
+    sampleRate?.takeIf { it > 0 }?.let {
+        "${(it / 100.0).roundToInt() / 10.0} kHz"
+    }
+
+fun FormatEntity.formattedFileSize(): String =
+    if (contentLength > 0) "${(contentLength / 1024.0 / 1024.0).roundToInt()} MB" else ""
